+++ test/test_integration.py
@@ -95,7 +95,7 @@ def latest_csv_columns(pattern):
 # Test functions
 def test_integration_with_large_tokamak_son(prepare_environment):
     """Test ACCERT with the large tokamak input."""
-    run_accert_and_check_output("tutorial/accert/LargeTokamak.son", "gold/output.fusion.out")
+    run_accert_and_check_output("tutorial/accert/LargeTokamak.son", "gold/output.large_tokamak.out")
     
     # For the large tokamak model, only check for the updated account CSV file.
     check_csv_files(["*_upd_acc_*.csv"])